styles for dropdown prop components and data moved to separate file
